# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Activity API",
        "version": "0.1"
    },
    "paths": {
        "/api/recomended": {
            "get": {
                "tags": [
                    "Activity"
                ],
                "summary": "Get a list of users recomendation",
                "operationId": "217f32774ba0cc72681f79856164ede0",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/activity": {
            "post": {
                "tags": [
                    "Activity"
                ],
                "summary": "Store Like or dislike",
                "operationId": "78adcc40d7f7370c0d2f7d5e5a53506a",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "from_profile_id": {
                                        "description": "id logged in user",
                                        "type": "integer"
                                    },
                                    "to_profile_id": {
                                        "description": "id liked by logged in user",
                                        "type": "integer"
                                    },
                                    "activity": {
                                        "description": "like or dislike",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/liked/{from_profile_id}": {
            "get": {
                "tags": [
                    "Activity"
                ],
                "summary": "Get a list of users liked",
                "operationId": "69e16416770fa3b04615235f50b91244",
                "parameters": [
                    {
                        "name": "from_profile_id",
                        "in": "path",
                        "description": "ID of user to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    }
}